Move code from submodule to init
